Use user-controlled `dbus_id` instead of nickname to register busname in DBus
DBus busname used to be `org.levak.Warfacebot.[realm].[nickname]`.
But, nickname can often contain characters that are not allowed in DBus core specs.
Now, the busname is `org.levak.Warfacebot.[realm].[dbus_id]`.

`dbus_id` is a CVar, user-controlled. When starting a bot as daemon (for dbus and
without wbd_launcher), it is now required to give it a `dbus_id` manually using
`-ddbus_id=xxx` in the command-line. In wbd_launcher, dbus_id is the associative key
used in bash.